📈 IMPROVE: Grid About - Change from grid to flex
