--- src/assets/styles/AboutImports.styl
@@ -1,4 +1,0 @@
-@import './Vars.styl'
-@import './Mixins.styl'
-@import './components/AboutFormat.styl'
-@import './components/GridAbout.styl'
