+++ src/assets/styles/components/AboutFormat.styl
@@ -1,64 +1,32 @@
-.gridAbout__item:nth-child(1)
-    grid-column 1/4
-    transition all 0.5s
-    aboutGrid()
-    .gridAbout__item--image
-        background-image url("/src/assets/static/img/Agustin-Blog.jpg")
-        background-position center
-        background-repeat no-repeat
-        background-size cover
-        aboutImage()
-    .gridAbout__item--specialty
-        justify-self flex-start
-        align-self center
-        grid-column 1/2
-        grid-row 1/2
-        transform rotate(-90deg) translateY(-0.5rem)
-        specialtyStyle()
+@import "../Vars.styl";
+@import "../Mixins.styl";
 
-// .gridAbout__item:nth-child(1).open
-//     grid-column 1/10
-//     z-index 100
-//     transition all 0.5s
+.gridAbout__item--container
+  position relative
+  width 33.33333%
+  height 100vh
+  display flex
+  overflow hidden
 
-.gridAbout__item:nth-child(2)
-    grid-column 4/8
-    aboutGrid()
-    .gridAbout__item--image
-        background-image url("/src/assets/static/img/mi_cara.jpg") 
-        background-position center
-        background-repeat no-repeat
-        background-size cover
-        aboutImage()
-    .gridAbout__item--specialty:nth-child(2)
-        align-self flex-start
-        justify-self flex-start
-        grid-column 1/4
-        grid-row 1/3
-        transform rotate(-90deg) translate(-6rem, -4.7rem)
-        specialtyStyle()
-    .gridAbout__item--specialty:nth-child(3)
-        justify-self flex-end
-        align-self center
-        grid-column 1/4
-        grid-row 10/13
-        transform rotate(-90deg) translate(-0.5rem, 6rem)
-        specialtyStyle()
+.gridAbout__image
+  width 100%
+  height 100%
+  object-fit cover
+  z-index 0
 
+.gridAbout__specialty--container
+  position absolute
+  display flex
+  flex-direction column
+  justify-content flex-end
+  transform rotate(-90deg) translateX(50%)
 
-.gridAbout__item:nth-child(3)
-    grid-column 8/11
-    aboutGrid()
-    .gridAbout__item--image
-        background-image url("/src/assets/static/img/Frame-hermanos H.jpg") 
-        background-position center 
-        background-repeat no-repeat
-        background-size cover
-        aboutImage()
-    .gridAbout__item--specialty
-        justify-self flex-end
-        align-self center
-        grid-column 1/4
-        grid-row 10/13
-        transform rotate(-90deg) translate(-1rem, 5.5rem)
-        specialtyStyle()
+.gridAbout__specialty
+  // align-self center
+  text-transform uppercase
+  color color8
+  font-size titlesFontSize + .2rem
+  z-index 10
+
+.gridAbout__specialty.second
+  display none
